13/05 update
Added volatility to the graph
Made the sound only play when a real transaction goes through because in
simulation mode it plays too often.

Added an additional check that the final sample of the EMA is actually
trending, so if you have say 2 or 3 EMA samples and the latest sample is
actually against the EMA trend it will spot the slight change in
direction and wait for the next sample to buy (if it is again back on
the trend). This makes the bot slightly smarter and really is a
necessity in case the EMA is still above your threshold but volatility
is suddenly high because of a sudden change in direction (I will add
option to turn that on/off in options later but I believe it should be a
default setting to on).